Fixing scan overrun when more than 64K samples are in the buffer.

+++ lib/mcc172.c
@@ -1833,7 +1833,7 @@ int mcc172_a_in_scan_status(uint8_t address, uint16_t* status,
 {
     struct mcc172ScanThreadInfo* info;
     uint16_t stat;
-    uint16_t buffer_depth;
+    uint32_t buffer_depth;
     bool hw_overrun;
     bool buffer_overrun;
     bool triggered;
@@ -1915,7 +1915,7 @@ int mcc172_a_in_scan_read(uint8_t address, uint16_t* status,
     struct timespec start_time;
     struct timespec current_time;
     uint16_t stat;
-    uint16_t buffer_depth;
+    uint32_t buffer_depth;
     bool hw_overrun;
     bool buffer_overrun;
     bool triggered;
